Fix overriding uploaders
Use prepend to avoid copying the entire files
