+++ config/initializers/image_uploader.rb
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+require_relative "../../lib/monkey_patching_image_uploader"
